fix build issues portaudio and cmake
Build portaudio without ASIO only as a start. I had to edit the
CMakeList.txt to be able to add a preprocessor makro that disables
the inclusion of ksguid.lib. I could not figure out how to do this
over the cli. We should likely create a patch for upstream and ask
them about advice.