Disable relocatable Qt with -no-feature-relocatable.
We build Qt with frameworks enabled. When frameworks are enabled, relocatable Qt
uses the framework bundle to determine what the relocatable path to Qt libraries
is:
https://code.qt.io/cgit/qt/qtbase.git/tree/src/corelib/global/qlibraryinfo.cpp?h=5.14.2#n537

We deploy standalone Qt dylibs without frameworks, so this causes a crash on
launch with Qt 5.14:
https://bugs.launchpad.net/mixxx/+bug/1871238

As a short-term fix, disable relocatable Qt. A longer term fix would be to build
with frameworks disabled, or to actually ship frameworks instead of standalone
libraries.